+++ substratevm/CHANGELOG.md
@@ -2,4 +2,8 @@
 
 This changelog summarizes major changes to Native Image, developed under the codename **Substrate VM**.
 
+## Version 22.0.0
+* (GR-33930) Decouple HostedOptionParser setup from classpath/modulepath scanning (use ServiceLoader for collecting options).
+* (GR-33504) Implement --add-reads for native-image and fix --add-opens error handling.
+
 ## Version 21.3.0